old-tests/webapps/XMLHttpRequest/tests/submissions/Opera/open-open-send.htm

<!DOCTYPE html>
<html>
  <head>
    <title>XMLHttpRequest: open() - open() - send()</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      var test = async_test()
      test.step(function() {
        var client = new XMLHttpRequest(),
            result = [],
            expected = [1,1,1,2,3,4]
        client.onreadystatechange = function() {
          test.step(function() {
            result.push(client.readyState)
            if(4 == client.readyState) {
              assert_array_equals(result, expected)
              test.done()
            }
          })
        }
        client.open("GET", "folder.txt")
        client.open("GET", "folder.txt")
        client.send(null)
      })
    </script>
  </body>
</html>
